split parameter definition into modules without requirements
